Refactor error handling to use custom exceptions
Several functions that previously returned errors as dictionary keys have been refactored to instead raise custom exceptions. A new file 'custom_exceptions.py' has been created to define these exceptions, which are UserNotFoundError and TokenNotFoundError. This change enhances readability, and makes error handling more effective and explicit in the codebase.
