+++ middleware/reset_token_queries.py
@@ -1,6 +1,8 @@
 from psycopg2.extensions import cursor as PgCursor
 from typing import Dict, Union
 
+from middleware.custom_exceptions import TokenNotFoundError
+
 
 def check_reset_token(cursor: PgCursor, token: str) -> Dict[str, Union[int, str]]:
     """
@@ -14,15 +16,14 @@ def check_reset_token(cursor: PgCursor, token: str) -> Dict[str, Union[int, str]
         f"select id, create_date, email from reset_tokens where token = '{token}'"
     )
     results = cursor.fetchall()
-    if len(results) > 0:
-        user_data = {
-            "id": results[0][0],
-            "create_date": results[0][1],
-            "email": results[0][2],
-        }
-        return user_data
-    else:
-        return {"error": "no match"}
+    if len(results) == 0:
+        raise TokenNotFoundError("The specified token was not found.")
+    return {
+        "id": results[0][0],
+        "create_date": results[0][1],
+        "email": results[0][2],
+    }
+
 
 
 def add_reset_token(cursor: PgCursor, email: str, token: str) -> None:
